refactor: async db helper to common
