+++ services/chat/pyproject.toml
@@ -41,7 +41,8 @@ dependencies = [
 [project.optional-dependencies]
 test = [
     "pytest",
-    "pytest-asyncio"
+    "pytest-asyncio",
+    "respx"
 ]
 
 [build-system]
