+++ services/office/models/__init__.py
@@ -8,19 +8,8 @@ from sqlalchemy import Text, func
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 from sqlmodel import Column, DateTime, Field, SQLModel
 
+from services.common import get_async_database_url
 from services.office.core.settings import get_settings
-
-
-# Create async engine for database operations
-def get_async_database_url(url: str) -> str:
-    """Convert database URL to async format."""
-    if url.startswith("postgresql://"):
-        return url.replace("postgresql://", "postgresql+asyncpg://")
-    elif url.startswith("sqlite://"):
-        return url.replace("sqlite://", "sqlite+aiosqlite://")
-    else:
-        return url
-
 
 engine = create_async_engine(
     get_async_database_url(get_settings().db_url_office),
