+++ services/office/pyproject.toml
@@ -47,7 +47,8 @@ dependencies = [
 [project.optional-dependencies]
 test = [
     "pytest",
-    "pytest-asyncio"
+    "pytest-asyncio",
+    "respx"
 ]
 
 [build-system]
